Move root `Select` components to their own folder
