+++ src/select/async-creatable-select.tsx
@@ -1,22 +1,22 @@
 import React, { forwardRef } from "react";
 import type { MutableRefObject, ReactElement, RefAttributes } from "react";
 import type { GroupBase, SelectInstance } from "react-select";
-import AsyncReactSelect from "react-select/async";
-import type { AsyncProps } from "react-select/async";
-import useChakraSelectProps from "./use-chakra-select-props";
+import AsyncCreatableReactSelect from "react-select/async-creatable";
+import type { AsyncCreatableProps } from "react-select/async-creatable";
+import useChakraSelectProps from "../use-chakra-select-props";
 
-export type AsyncSelectComponent = <
+export type AsyncCreatableSelectComponent = <
   Option = unknown,
   IsMulti extends boolean = false,
   Group extends GroupBase<Option> = GroupBase<Option>
 >(
-  props: AsyncProps<Option, IsMulti, Group> &
+  props: AsyncCreatableProps<Option, IsMulti, Group> &
     RefAttributes<SelectInstance<Option, IsMulti, Group>>
 ) => ReactElement;
 
-const AsyncSelect = forwardRef(
+const AsyncCreatableSelect = forwardRef(
   <Option, IsMulti extends boolean, Group extends GroupBase<Option>>(
-    props: AsyncProps<Option, IsMulti, Group>,
+    props: AsyncCreatableProps<Option, IsMulti, Group>,
     ref:
       | ((instance: SelectInstance<Option, IsMulti, Group> | null) => void)
       | MutableRefObject<SelectInstance<Option, IsMulti, Group> | null>
@@ -24,8 +24,8 @@ const AsyncSelect = forwardRef(
   ) => {
     const chakraSelectProps = useChakraSelectProps(props);
 
-    return <AsyncReactSelect ref={ref} {...chakraSelectProps} />;
+    return <AsyncCreatableReactSelect ref={ref} {...chakraSelectProps} />;
   }
-) as AsyncSelectComponent;
+) as AsyncCreatableSelectComponent;
 
-export default AsyncSelect;
+export default AsyncCreatableSelect;
