+++ src/select/async-select.tsx
@@ -1,22 +1,22 @@
 import React, { forwardRef } from "react";
 import type { MutableRefObject, ReactElement, RefAttributes } from "react";
 import type { GroupBase, SelectInstance } from "react-select";
-import AsyncCreatableReactSelect from "react-select/async-creatable";
-import type { AsyncCreatableProps } from "react-select/async-creatable";
-import useChakraSelectProps from "./use-chakra-select-props";
+import AsyncReactSelect from "react-select/async";
+import type { AsyncProps } from "react-select/async";
+import useChakraSelectProps from "../use-chakra-select-props";
 
-export type AsyncCreatableSelectComponent = <
+export type AsyncSelectComponent = <
   Option = unknown,
   IsMulti extends boolean = false,
   Group extends GroupBase<Option> = GroupBase<Option>
 >(
-  props: AsyncCreatableProps<Option, IsMulti, Group> &
+  props: AsyncProps<Option, IsMulti, Group> &
     RefAttributes<SelectInstance<Option, IsMulti, Group>>
 ) => ReactElement;
 
-const AsyncCreatableSelect = forwardRef(
+const AsyncSelect = forwardRef(
   <Option, IsMulti extends boolean, Group extends GroupBase<Option>>(
-    props: AsyncCreatableProps<Option, IsMulti, Group>,
+    props: AsyncProps<Option, IsMulti, Group>,
     ref:
       | ((instance: SelectInstance<Option, IsMulti, Group> | null) => void)
       | MutableRefObject<SelectInstance<Option, IsMulti, Group> | null>
@@ -24,8 +24,8 @@ const AsyncCreatableSelect = forwardRef(
   ) => {
     const chakraSelectProps = useChakraSelectProps(props);
 
-    return <AsyncCreatableReactSelect ref={ref} {...chakraSelectProps} />;
+    return <AsyncReactSelect ref={ref} {...chakraSelectProps} />;
   }
-) as AsyncCreatableSelectComponent;
+) as AsyncSelectComponent;
 
-export default AsyncCreatableSelect;
+export default AsyncSelect;
